Split `Lang.Crucible.JVM.Generator` into two separate modules.
* Lang.Crucible.JVM.Context contains the JVMContext datatype.

* Lang.Crucible.JVM.Translation.Monad contains the JVMGenerator monad.

Module Lang.Crucible.JVM.Translation.Monad is considered to contain
internals only relevant to Lang.Crucible.JVM.Translation, so we don't
need to re-export this part from the top-level Lang.Crucible.JVM.
