+++ crucible-jvm/crucible-jvm.cabal
@@ -76,9 +76,10 @@ library
   exposed-modules:
     Lang.Crucible.JVM
     Lang.Crucible.JVM.Types
-    Lang.Crucible.JVM.Generator
+    Lang.Crucible.JVM.Context
     Lang.Crucible.JVM.Simulate
     Lang.Crucible.JVM.Translation
+    Lang.Crucible.JVM.Translation.Monad
     Lang.Crucible.JVM.Translation.Numeric
     Lang.Crucible.JVM.ClassRefs
     Lang.Crucible.JVM.Class
