+++ crucible-jvm/src/Lang/Crucible/JVM/Simulate.hs
@@ -71,9 +71,10 @@ import           What4.Utils.MonadST (liftST)
 
 -- crucible-jvm
 import           Lang.Crucible.JVM.Types
+import           Lang.Crucible.JVM.Context
 import           Lang.Crucible.JVM.Translation (translateMethod, jvmIntrinsicTypes, jvmExtensionImpl)
 import           Lang.Crucible.JVM.ClassRefs
-import           Lang.Crucible.JVM.Generator
+import           Lang.Crucible.JVM.Translation.Monad
 import           Lang.Crucible.JVM.Class
 import           Lang.Crucible.JVM.Overrides
