+++ crucible-jvm/src/Lang/Crucible/JVM/Translation.hs
@@ -61,8 +61,9 @@ import qualified What4.Partial.AssertionTree as W4AT
 
 -- crucible-jvm
 import           Lang.Crucible.JVM.Types
+import           Lang.Crucible.JVM.Context
 import           Lang.Crucible.JVM.Translation.Numeric
-import           Lang.Crucible.JVM.Generator
+import           Lang.Crucible.JVM.Translation.Monad
 import           Lang.Crucible.JVM.Class
 
 import           Debug.Trace
